Revert "Feat/issue49"
