--- client/src/components/Nav.css
@@ -1,50 +1,0 @@
-*{
-    text-decoration: none;
-    margin: 0;
-    box-sizing: border-box;
-}
-aside {
-    background-color: #0C6170;
-    width: 230px;
-    height: 100vh;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    padding: 0;
-
-    
-
-    
-}
-
-aside h1 {
-    color: #fff;
-    border-bottom: 1px solid #796767;
-    margin: 0;
-    padding: 50px;
-    padding-top: 30px;
-    padding-bottom: 30px;
-    display: flex;
-    justify-content: flex-start;
-    gap: 10px;
-}
-
-nav {
-    display: flex;
-    flex-direction: column;
-    padding: 50px
-    
-}
-
-.tag-link {
-    color: #fff;
-    margin-bottom: 20px;
-    display: flex;
-    gap: 15px;
-
-}
-
-
-
-
-
