--- client/src/pages/Home/Home.css
@@ -1,38 +1,0 @@
-*{
-    text-decoration: none;
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-.container-home {
-    display: flex;
-
-}
-
-.box-right-home {
-    width: 100%;
-    
-}
-.box-header-home {
-    padding: 35px 0;
-    /* background-color: red; */
-    border-bottom: 1px solid #796767;
-    height: 93px;
-    display: flex;
-    justify-content: flex-end;
-    padding-right: 30px;
-}
-
-.box-image-home {
-    display: flex;
-    justify-content: flex-end;
-
-}
-
-.box-image-home img {
-    margin-top: 100px;
-    margin-right: 100px;
-    width: 50%;
-
-}
